Instead of the React context and memoize approach, I decided to fully utilize re-frames signal graph to not only extract, but also compute relevant results in query functions. Computation will not re-evaluate unless the input signals they depend on changed, so it is like memoize in a way, but without having to clutter my view functions or to create awkward abstractions.
I also managed to implement the variable order selection this way and it works
quite well.

The only issue is now that values do get recomputed when the componts that
subscribed to them are unmounted and mounted again, which happens when I switch
to a different mode. Maybe I can create a common component to initialize all
computed values, but they may not be needed, so it feels like a waste of
resources. Maybe I’ll have to go back to caching those results in the db.