I found that with my previous approach, values got recomputed twice if the input expression changed. This was because the `:value` sub/computation depended on `:expr` and `:varorder` subs, which would fire 2 separate signals when the `:changed-formula` event updates the db with a different expression and a different varorder. I solved this issue by creating an additional “bundled” sub, which would then only fire once if either expr or varorder or both change in the same event.
There is still the redundant call to `expr/find-vars` in `:varorder-permutations`
sub, because we already have the varorder and only need to sort it, which is
much more efficient than having to traverse the expression tree again. But
I cannot have it depend on `:varorder` sub since then it would recompute when
a user selects any other varorder. I tried to solve that by providing the sorted
varorder from the component to the input of the sub, but this causes the sub
to be recomputed twice, because in the downward signal-flow the old value for
varorder is unchanged and after the view node is updated with the new varorder,
`:varorder-permutations` will be called again with that new value. I might have
to somehow store the sorted varorder or make it a derived computation.