I rebuilt the split-view from my previous approach using the DevbookHQ/splitter library for React. Changes in one pane are sync’ed with the other. Function menu does not work yet since it requires knowing to which view a function should be applied to. Should also be normal buttons instead of radio buttons.
Maybe view selection should be on focus or with a switch or directly in the
menu? This is the main question right now.

Added flow-storm debugger since it now works with Helix components and could
be useful in some cases. Instrumentation seems to work only with namespaces,
not single functions right now. Removing instrumentation only works on page
reload.