Added split-view controls to change from single view to split views, split orientation or swap the views. I got rid of the `:active?` entries in the db `:views`, since there is no need for views to be “active” – we can just deactivate the split mode and have the first view be the default for single view mode (swapping enables to change that). This way, there will always be two views in the db, where the second one may just not be presented.
As for now, views are not cached, so the second view will always
calculate the selected function again if we switch back and forth from
single to split views. This is not ideal, but I will keep it like that
for now since I am not sure how much caching should be done and what
should be cached. I could just hide the HTML element of the second view,
but that may blow up the memory usage in some cases and would always
refresh both views on input change even though one is not visible.