Changed F-Selfi to static CA with formDNA input, added dna sub
A static CA does make more sense for FORM tricorder, since it is used more as a
toolbox for quick operations and lookups instead of deep, focussed work. So
results have to be more immediate and useful at a glance.

The performance overhead of rendering a CA evolution window at once is minimal
with the level of detail required in most cases and it is simpler to switch
between functions without animation state floating around.

I may add a “pagination” feature to advance/regress the evolution window, since
in some cases patterns may only emerge after many generations. Maybe the longer
history can then be knitted together for image export.