app-db model changes, performance (re-renders) & CSS improvements
Decoupled the app-db representation of `views` from frame data, which manages
how to display/arrange multiple windows inside the `OutputArea` frame (think
Emacs frames and windows). Views data structure didn’t change at all, but it now
represents only which function components should be rendered and their display
order. The frame data decides the layout orientation and how many windows should
be displayed, but says nothing about their contents. This separation saves many
unnecessary re-renders.

The drawback to this approach is that the window count in `:frame` is now
coupled to the vector size of `:views`, which can lead to a mismatch when
re-frame/refx state updates are not completely in sync with component updates in
React (see `:view` sub). Not sure if there is anything I can do about it, except
for putting some assertions in the event and subscription handlers or provide
fallback mechanisms. Maybe I still need a better data model or a better way to
think about this.

I also decided to move re-frame/refx requests deeper into the component tree so
that app-db state is only pulled and pushed where it is actually being used (or
rather: where it conceptually makes sense, but I still try to figure that out),
which also saves some unnecessary re-renders.

`ViewPane`s are now wrapped inside a div, which prevents them from re-rendering
when the size of the `Splitter` changes, which took me some annoying debugging
sessionk to find out.

Also made some CSS changes that makes the `OutputArea` only use up the rest of
the browser window height and have the `ViewPane`s scroll inside their own area
on overflow.