Cache in app-db, refactored & cached F-Selfi
I observed that `F-Selfi` was slowed down significantly by the initialization of
`evolution` data. I could interleave it with the actual painting of the canvas,
but that would make the data inflexible and mix separate concerns, which I would
like to avoid.

Reacts `useMemo` is useless here since the function components get unmounted
each time windows are swapped or different tabs/modes are selected. Keeping the
DOM elements on the page is a non-option here, because that would clutter the
DOM significantly when lots of function components have been mounted.

I also don’t know if I could do this with Radix’s tabs component, since the tab
contents get unmounted on deselect. Window swapping may be done with CSS
instead, but I would have to decouple the window order from the views
vector (which might actually be a good idea).

Anyways, my solution is to cache the `evolution` data in re-frames app-db and
retrieve it from there after it has been computed the first time in
`F-Selfi` (where an event gets dispatched to update the cache).

The cache needs to be invalidated every time a dependency from the resp. cache
item has been changed, i.e. when an input event happens. The input event defines
its own set of dependencies that it requires the invalidation event to check, so
I have to always make sure that these line up correctly to prevent stale cache
data from lingering around.

Maybe a better solution will come along some day, but I think this is a good
compromise. For now, I’ll only cache data that is expensive to compute, like CA
data and maybe calculation results to keep complexity low.