Exchanged refx for re-frame and created adapter for helix
Tested this whole setup and confirmed that it works just like before with refx.

I have to watch out for potential side-effects that might occur due to the
adapter code, which was taken from
https://github.com/pitch-io/uix/blob/master/docs/interop-with-reagent.md#syncing-with-ratoms-and-re-frame
and modified to work with Helix instead of UIx.

I guess that the only thing to watch out for is where subscriptions happen in
the components and maybe also dispatches, but the pure re-frame handlers should
not be affected by the migration.

One of the motivations behind the switch is that there are some newer features
and maybe fixes in re-frame that may not become merged into refx any time soon
due to it being a modified copy of re-frame instead of a “clean” fork and having
potentially incompatible changes in refx. Also see
https://github.com/ferdinand-beyer/refx/issues/6 where the creator talks about
getting closer to re-frame again, but it¨s not really clear if he is actively
working on it and I do not want to depend on it at this time, especially since
the project is still in an alpha state.

I hope that re-frame will at some point become independent of reagent, but when
this time comes I have already made sure that I’ll merely have to remove the
adapter code and the var names or any other code doesn’t need to be changed.