Event-level error handling, error display, submit mode on input
Working on error handling for formula input. Re-frame has a nice
`reg-event-error-handler` that catches unhandled exceptions thrown in the
interceptor/handler chain. I used this to dispatch an event that stores the
error in app-db, so I can react to it from the UI. Afterwards, the `clear-error`
interceptor will remove it whenever a new input event happens.

Maybe later on this can be differentiated by what kind of error it is or which
components are affected (should atm only be formula input unless there is a
bug).

However, I am working on a new input editor built on codemirror that helps the
user to avoid unbalanced parentheses and other common syntax errors from the
start. I found that updating the app-db on change from the input is actually
quite convenient and works well in most cases, unless many variables are
involved. A new “submit mode” can be toggled to only update on submit, but I’m
not sure how to best communicate that through the UI.