Rewrote FormulaInput to use CodeMirror
The editor is initialized on-mount (`useEffect` hook) by creating the
`EditorState` and instantiating the `EditorView` with that state and the DOM
reference (via `useRef` hook) as a parent.

I added some useful extensions such as `bracketMatching` to match bracket pairs
and `closeBrackets` to auto-close brackets and some of the things recommended
for a minimal editor.

The theme is still wip, but there is a distinction between light- and darkmode.
Having the editor recognize the `appearance` state for darkmode from Helix was a
bit tricky, because the actual darkmode setting within the editor is a “facet”,
which is a static value that cannot be mutated from anywhere and has to be
changed indirectly via a “statefield”.

First, the darkmode state from Helix has to be recreated and synced as a
`StateField` within the editor. When `appearance` changes, a `useEffect` hook
dispatches a transaction that updates the editor state by passing a
`StateEffect` for the new value. This effect will then be recognized in the
`update` method of the statefield, where its value gets returned as the new
value for that field. This new value can then (as an extension) set the
`darkTheme` facet for the view, which causes the DOM element to use the `dark`
class for the CSS. This seems like a huge amount of machinery for such a simple
task, but maybe there is an easier way that I missed.

Something similar had to be done for the `submit-mode` state, except here my
custom `submit-keymap` calls a command (when the user types shift-enter) that
extracts the statefield. If its value is true, the editor input gets sent to
app-db. I could not use the Helix state directly here, because the function
doesn’t react to its updates. I also have to use that statefield in the
`updateListener` extension to not send input when submit-mode is active.

Maybe I can remove the component-local `code` state since there is no point in
having it when CodeMirror manages the same state internally.