Rewrote most output components from Helix to web-components
This is an important step for multiple reasons:

These components are not tied to form-tricorder or any web framework at all and
it should be possible to use them independently by just loading a js file and
write them as custom HTML tags.

Using web-components allows for better encapsulation, keeps CSS styles
consistent and avoids interference with other elements on a website/app. Still,
reactivity and interactive elements can be used just like with React.

I still have to figure out a lot of things about how this works, but the zero
library makes it already pretty easy to define web-components and enabled me to
rewrite all significant output components.

`vmap` is mostly done, but one could imagine some interactive elements there for
filtering of values.

`fgraph` now produces all graph-like visualizations for expressions, but at some
point I want to rewrite the D3 code from `form-svg` and integrate it more
directly with the web-component.

`selfi` of course is far from finished and I had to skip the caching for now,
because I don’t know how to handle it in a clean way and also because it would
need to dispatch to re-frame, so I may have to pass an event handler to not
break encapsulation.

`vtable` is mostly done, but I had to use a different styling approach that does
not depend on my global setup with Stitches and does not output a static CSS
file (users should not have to import more than the js). The cljs library
“garden” seems to work well here, since it just converts Clojure data structures
to a CSS string, which can be attached to the shadow root as a `CSSStyleSheet`
via zeros `::z/css` property. However, I still have to figure out how to best
enable customization of CSS properties and how far this should go.

I’ll migrate all web-components to a new library very soon, so form-tricorder
can just use that as a dependency and I can include these components in my Clerk
notebooks (hopefully), in Observable or in future apps.