Recreated vtable styling per CSS argument
Using the Garden library for now to create CSS styles as a string that can be
easily passed to a web-component and extend the stylesheet of its shadow-DOM.

CSS variables can be used to integrate with my Stitches config. Not sure yet if
I’ll keep Stitches or Garden, but it works for now.