common ui components, redesigned input sync & varorder select
Added the shadcn/ui components which I have adapted in my other repo to Stitches
+ Helix. Now these components can be reused and easily customized across the
app (and possibly future apps). Testing the components in the new `foobar`
namespace.

The color variables now match the new scales and I harmonized the semantic
names. I realized that I need each of the common components (and their variants)
to have an “outer” and an “inner” theme, to be usable on both backgrounds in the
app. The code duplication and extra prop here isn’t ideal, but I don’t have a
better solution yet. Maybe it is the price to pay for having different layers in
the UI.

The formula input has been redesigned to better represent the sync state and how
to toggle it on or off. When toggled off, an “enter” button appears to apply the
formula manually (otherwise, shift+enter also does that).

The varorder select has been rebuilt on top of the new common select component
and has a new home in the bottom-left corner of the app, which can be collapsed
to not obstruct the view. This lets users change it globally instead
of (re)mounting the same component in different view-panes.

I also refined the top toolbar which looks a lot better now.

As for dev tooling, I switched from yarn to pnpm as a package manager.
Unfortunately, had to use `shamefully-hoist = true` since the flat file
structure is a problem for shadow-cljs (had a similar issue with yarn, so not
much of an improvement).

Not sure yet if I want to stick with Radix or lucide-react icons, but eventually
I’ll just copy the svg later on to remove these dependencies altogether.