Implemented new color scale approach, mapped old to new variables
I added a step 0 (white) and 31 (black) to my final scale parameters to extend
the max range when needed (and still work across light/dark modes).

I also added the coral/signal scale. Chroma and hue parameters for all scales
may still be adjusted and fine-tuned in the future as they are not ideal yet.

Values will be hard-coded later on to avoid scale generation and the apcach
dependency at runtime. I am also thinking about switching to shadow-css from
stitches at some point in the future for even better performance.

Semantic colors are currently set to scale values that correspond to the old
mappings, but will have to be adjusted, harmonized and fine-tuned later on.