formDNA function for eval mode
Splitting the dna into groups of 4 makes it easy to visually organize it by
resizing the browser or view pane, so there is no need for further grouping.

I added a select component to choose an encoding as digits in NUIM and NMUI
format. However, the constants format still depends on some order, which is NUIM
by default. I will have to make this clearer in the UI, but I guess that I’ll be
the only user of this function anyway, so it’s not a priority.