migration to shadow-css: button, toggle, radio-group, formula-input
I got reminded of the fact that CSS classes do not apply in the order defined on
the element, but in their order in the CSS file. This wasn’t an issue with
Stitches, but now with static CSS I have to be aware that I cannot declare style
overrides on components except with `style` attributes.

With my variant system, I must not have any intersection of css properties
between the base class and all variant groups. This is a minor inconvenience
which may become more annoying later, but I am not sure if things like “Class
Variance Authority” (which people seem to use with Tailwind) work with
shadow-css or if I want all that complexity.