refactored ExportDialog & vmap fn, colors use light-dark() CSS fn
I heavily refactored the `ExportDialog` component to become generic and modular,
independent of the specifics of the exported thing and the logic and composition
of export options. It is a much cleaner, disentangled design and I can flexibly
assemble it for each visualization.

Also refactored the vmap function components to better separate vmap and vmap
perspectives and to build the specific export dialog for them with some new
options. I also adapted them to the new interface from `formform-vis`.

Changed CSS output of light-/darkmode colors to use the `light-dark()` CSS
function in conjunction with `color-scheme` instead of the
`prefers-color-scheme` + `:root[data-theme=”dark”]` solution for the mode
switch. This is much simpler, less repetitive and allows me to set modes
explicitly anywhere in the cascade (which works nicely in the `ExportPreview`
component).