changed url params update behaviour, tooltip cmp, mode & fn keybinds
URL params now only update when the user clicks “copy link” in the menubar,
instead of updating on every significant input. This may speed up performance
and avoid state corruption due to incomplete updates or errors. It also makes
sharing the URL easier so people don’t have to remove search params. The button
acts as a save button that also copies the URL after params have been set, so it
is still as or even more convenient than before.

I implemented a tooltip component on top of the radix primitive to give feedback
about the copied URL.

The function menu now has keyboard shortcuts for every mode and every function
that work similar to keybindings in Emacs and are mnemonic. All of them have a
control-key prefix and you first select modes, then functions.

Currently the mode keybind also opens the submenu of the specific mode. I might
change that to show the available functions for that mode in a popup somewhere
below instead, which would be less noisy and avoid focussing the menu when the
user is in formula input.