common varorder defaults, graph-notation styles, renamed to circle
“Graph notation” has been renamed to “Circle notation” in the UI (unsure if I
should use that name internally too), since it’s much clearer to common users.

I reimplemented the `styleclass` property of the graph visualization from
formform-vis with an ad-hoc radio-button UI. It is persisted via query params in
the url as `graph-style` (might have to rename to `circle-style` for
consistency).

In the app-db, I stored it under `:modes` → `:expr` → `:graph-style`. Other mode
options can be stored in `:modes` as well, but need not necessarily be persisted
in the url if that doesn’t make sense. I am unsure yet if graph-style should,
however I believe that it is something I would like to be shareable since it
influences the way I think about a given FORM.

Also fixed some smaller bugs.