fixed formula input losing local state + CSS adjustments & fixes
Fixed formula input to pick up formula from app-db in its local state on init

Added some `:focus-visible` rings and fixed visibility issues.

Added a border on `FuncOpts` component with a label icon to better separate
controls from the visualization and give the user a hint on its purpose. Not
sure if I keep it though.

Removed the shadow on the function menu buttons to better fit the visual style
of the app.