value filter component, filtered subs, filter params in app-db
Filter query params are not tested yet, I still need to generate them for the
url on copy link, but the filters seem to work nonetheless.

The value filter component is mostly done, but needs a better layout.

I added UI colors for all 4 value constants, which makes the value toggle
buttons more easily recognizable.

formDNA is now computed from the `:input/->results` sub, because it is more
straightforward to extract the dna from the results vector, which also provides
the vspace needed for filtering. Now that dna is not reversed anymore (which I
still need to implement in formform so it’ll remain reversed for the time
being), I don’t need to worry about performance implications.

Also added toggle-group style variants, which include a nice vmap-shaped toggle
arrangement for the interpretation term filters.