added value filters for dna fn., fixed set ops to be radio buttons
I added the value filters to formDNA function. This isn’t really that useful in
most cases, but I like how it enables newcomers to build intuition about how
formDNA is structured as a code and maybe there will be some use-case for
incomplete formDNA with “holes” in the future.

Set operation selection has been with a single-selection toggle group, which had
the side-effect of none selection being valid (throwing an error), so I had to
adjust the radio-group component to apply the same toggle-group styles and still
default to bullets.

Not sure yet if this is a good approach, since the visual distinction between
radio buttons and toggles now vanishes, but I guess in this special case it’s
fine. Maybe I can add more visual cues later on.