value-filter collapsible/toggle, stays open when filters changed
The filters are collapsed by default as long as there are no changes to them, so
that users can work undistracted. However, when changes are made, the filters
should stay open so that it is immediately apparent that values are filtered.
This is important as the output may be confusing users otherwise.