CA webc state-loss & memory leak fix, prod. build setup, etc.
`:input/→ca-x` subs now return only a `CASpec` and the actual automaton object
gets created directly in the `F-Automaton` component. This allows for more
predictable updating behavior in re-frame and React, because we are computing
and passing immutable data instead of a state machine and the automaton will be
tightly coupled to the web component anyway.

The CA function components now skip the initial render when they change the key
prop of `F-Automaton`. to prevent accidental unmounts.

I rewrote `F-Automaton` , because my previous approach to declare the web
component directly in the component markup was causing lots of (component
crashing) issues with state loss and memory leaks due to detached references,
especially when using `StrictMode` in React. What seems to work best is to
imperatively create the web component directly in `use-effect` and append it to
an empty container, using a ref to update its props later on.

From my tests this has been very robust so far, but still leads to occasional
state loss in the web component when the input changes in incompatible ways.
However, these errors don’t lead to noticable issues and without StrictMode seem
to only happen on split view changes (may be related to the cleanup function,
but no cleanup leads to detached references - I’ll have to investigate further).

Other changes:

- production build setup that has been tested successfully
  - es2018 is needed by subzero, because it seems to use some newer feature

- changed app entry to `init-fn` instead of `entries` in shadow-cljs.edn

- bumped radix and codemirror deps

- added re-frame-10x to dev for better debugging of re-frame stuff

- `ExportDialog` now clears the timeout from `on-export` handler to prevent a
memory leak

- disabled shift-key handling in `onSelect` of `FunctionMenu` component for now,
since `window.event.shiftKey` is not well supported in browsers. I am thinking
about a different approach that will also be more accessible without a keyboard.

Note: a lot of changes in the diff are from reformatting code indentation, which
happened when I inserted println statements in various components for debugging.