added split-view buttons to function menu, fixed keybind handler
Redesigned the `FunctionMenu` component to show a button for split-view on each
menu item. I think this is the most accessible and understandable solution.

Moved the menu keybind handler and `keybind-mode-value` state to `FunctionMenu`
so it wouldn’t trigger a rerender of the whole app on keypress. Also fixed the
handler itself to not close a submenu when ctrl is pressed again while it is
open.

Uncommented the `:after-load` line in shadow-cljs.edn, because otherwise it
causes issues with shadow-css on reload.