refactored export, system-color-scheme, formform 0.2.0 adjustments
I refactored the export dialog component from vmap export to work with graph
visualization export as well, separating it into common components that can be
put together as needed. It’s not a perfect abstraction, but it is good enough
for now. The appearance options in graph export rerender the whole visualization
instead of only the relevant parts, which is suboptimal, but I may postpone
fixing that until I get to rewriting the old JS/D3 implementation.

I noticed that the automatic recognition of color scheme preferences by the user
in CSS doesn’t apply to the CodeMirror input field as well as the various graph
visualizations I built with D3 a long time ago. The solution was simple with
CodeMirror, but for D3 I had to create a `system-color-scheme` state in app-db
which gets updated when the user changes their preference. I only really use
that when the `light-dark()` function from CSS is not an option, but it may be
beneficial in future use-cases.

Since I made some significant API changes in formform 0.2.0, I had to make some
adjustments in tricorder as well.

The most pervasive one was the case change in constants and formDNA from
uppercase to lowercase, since this relates to many helper functions as well that
don’t directly interface with formform.

Another big adjustment was that formDNA now reads from left-to-right instead of
right-to-left. Fortunately, this removes many reversing operations but I had to
carefully watch for implicit assumptions of directionality in the code.

Since formform api functions now consistently take the options map last this had
to be adjusted as well.

Lastly, of course, there were many changes in the formform.emul api.