lots of changes, mostly config panel for emul mode, 1d + 2d slider
I built a config panel for CA functions in emulation mode, on which I wanted to
include most of the customizations that are available in formform.emul and
provide a nice UI for them.

For some options I needed a slider, so I adapted the slider component from
shadcn/radix which works well. However, there is no 2d slider so I had to create
my own to easily adjust the 2d ini figure position. Luckily, someone actually
wrote a simple implementation of the base functionality in Clojure, which I just
had to adapt to Helix and customize/extend for my own needs + some styling.

The config panel is close to being finished, but URL params are still missing.